Working on getting KeystoneJS integrated
The project runs, but I cannot access the admin panel. Keystone is running its own server independent of the express server which is on a different port, so this is an interesting situation I need to investigate further. I set keystone to run on port 3001, so localhost:3001/keystone should load the admin panel, but I get an error from the express server. I presume this is a routing related issue combined with the port issue. I will investigate this further Friday.