The word search game is more or less functional now
Now the admin (only admin) can create a new game, select word search and fill in the form and the word search game will be generated. It seems to be saving and loading from the database with no issues at this point. You can click/touch 2 points to select a letter range and if found, I mark the answer as found. It still needs work to make the play experience more smooth, i.e. automatically deselecting words when clicking somewhere else on the board and leaving the word you found crossed on the game board. I also would like to indicate the player is a "winner" if they find all the words.