seperated game template controllers from main controller
It should be easier to create game templates now, I can create seperate controllers and seperate schemas for each game. The game veiw and create game veiws are still merged, but the logic in those is minimal enough, I am going to leave it that way. It will only become an issue if there are too many game templates, which is unlikely to happen anytime soon.